Split githubrouter
